Move Event to model
