+++ app/src/main/java/com/oxapps/materialcountdown/db/EventDao.kt
@@ -6,6 +6,7 @@ import android.arch.persistence.room.Delete
 import android.arch.persistence.room.Insert
 import android.arch.persistence.room.OnConflictStrategy.REPLACE
 import android.arch.persistence.room.Query
+import com.oxapps.materialcountdown.model.Event
 
 
 @Dao
